Two displays
One can now load both an aerial image and a camera and apply points to each independently.

Add text labels to read out the number of points selected in each image

Wrote .list_points() method for HomographyView objects. This should be called to retrieve a list of selected points (in correct order) in image coordinates. This is used for updating the status message and should be used for computing homographies.