Added custom class for showing homography results
Wrote HomographyResultView class for the middle pane of the homography GUI. This will display the "goodness" of the homography via an image.

Connected this class to the middle pane in QT designer

Added a status message showing the purpose of the middle pane until an image is loaded into it.