Wizard GUI for new project creation
Outlined GUI for inputting new project information.

Added this to app.py so ctrl o/open project opens this dialog.

This dialog opens as a modal, so the user may not interact with the GUI until this box is closed.

The browse buttons have been connected to file dialog boxes.