Remove framerate field from GUI
Summary:
Removed any instance of framerate in the UI and config file, since
FPS is being calculated implicitly by the server.

In app.py, I commented out lines that called CloudWizard.configFiles
since, it was throwing a syntax error which did not allow me to run app.py

Tested: I got a screenshot of the new UI without framerate, farmerate is
not created in the config.cfg file, and I searched for all other instances of
video_fps and delted them from the repo.