Fixed VideoPlayer end of video stop bug
Summary:
The QMediaPlayer did not show a EndOfMedia status IMMEDIEATELY  when the media has played
Now, the player is stopped when the slider duration hits the end

Resolves issue #94

Tests:
Playing the feature video from the GUI til the end displays the new
behavior.  Before, it would end with a white screen, and would not
immediately go to a stopped state when the video had ended.